Implementation of Proof::Update
This method is useful for cached proofs, it can take a valid proof and
update it with all of a block UTXOs, yielding a new (also valid) proof.

It replaces all intermediate values in a proof with the new ones, also
adds new positions that where't in the original one. Finally, calculates
the new positions for each target.